Added a circbuf_used_buf method. Changed safety checks.
Our safety checks were using the circbuf_avail_buf method
to see if we had sufficient data, however that method is not
suitable, as it returns the _available_ space and not the
utilized bytes. So it was returning the free space...

This check prevents us from reading garbage data from the buffers.