Adding mode to emit metrics over binary protocol
To complement the binary frontend, there is now a binary back end,
so that sinks can recieve the metrics in a binary form. This method
has no pre-conceived notions of how to build up the keys, and instead
just provides the "value type" (sum, count, min, etc), as a flag,
instead of modifying the key itself. The key is now passed directly
through.