Log to stderr when not daemonizing
The previous behavior was to log to stderr iff stdout was a TTY.
However, this made it cumbersome to log in environments where syslog is
not available (Docker) or stdout/stderr logging is conventional
(Supervisord), since it may not always be trivial to arrange for a
pseudo-tty to be allocated for no purpose other than persuading statsite
to log to stderr.

In cases where statsite is daemonizing, logging to stderr has little
value since daemonization closes stderr. When statsite is not
daemonizing, it's likely that stderr logging is desirable, and when not,
trivially disabled by redirecting stderr to /dev/null.