HTTP+JSON serialization sink
In order to reduce pull request size and interdependence,
here is an incremental step for a stdout JSON output sink. The stdout
writes will of course be replaced with a lifoq push for libcurl posting.

The current sink infrastructure is a bit hard to unit test - once the
HTTP plumbing is in place, there is an integration test for it.