Add udp_rcvbuf configuration option to support larger socket buffers
This adds a configuration option to the Statsite config file called
`udp_rcvbuf` which is used to make a setsockopt() call on the UDP
listen socket to set the SO_RCVBUF option.

Busy UDP sockets can benefit from having a larger than normal socket buffer
compared to the rmem_default setting.  This allows one to instruct
Statsite to set a different size buffer for the UDP socket.  Presumably,
rmem_max.