Simpler extended metrics counters (only add rate)
The actual extended metrics on counters doesn't seem to make a lot of
sense. They are implemented to be the same one that the timers have, but
most of the generated metrics on the timers are not really that useful
for counters. Also, the current metrics are a little confusing due to
the meaning of count (which is normally the sum on others statsd
implementations).

The only metric which seems relevant for the counters is the rate, so
the approach on this commit is to directly drop the extended counters
include config, as the extended counters only would add the rate. Also,
the meaning of count is changed to be the sum of the metrics received.

Co-authored-by: Ivan Izaguirre <ivanizag@gmail.com>
Co-authored-by: Taras Sotnikov <tsotnikov@tuenti.com>